move tests to libcollectionstest
